refactor: move bridge.util and types to shared directory
