+++ shared/modules/bridge-utils/quote.ts
@@ -1,0 +1,36 @@
+import { QuoteRequest } from '../../types/bridge';
+
+export const isValidQuoteRequest = (
+  partialRequest: Partial<QuoteRequest>,
+  requireAmount = true,
+): partialRequest is QuoteRequest => {
+  const STRING_FIELDS = ['srcTokenAddress', 'destTokenAddress'];
+  if (requireAmount) {
+    STRING_FIELDS.push('srcTokenAmount');
+  }
+  const NUMBER_FIELDS = ['srcChainId', 'destChainId', 'slippage'];
+
+  return (
+    STRING_FIELDS.every(
+      (field) =>
+        field in partialRequest &&
+        typeof partialRequest[field as keyof typeof partialRequest] ===
+          'string' &&
+        partialRequest[field as keyof typeof partialRequest] !== undefined &&
+        partialRequest[field as keyof typeof partialRequest] !== '' &&
+        partialRequest[field as keyof typeof partialRequest] !== null,
+    ) &&
+    NUMBER_FIELDS.every(
+      (field) =>
+        field in partialRequest &&
+        typeof partialRequest[field as keyof typeof partialRequest] ===
+          'number' &&
+        partialRequest[field as keyof typeof partialRequest] !== undefined &&
+        !isNaN(Number(partialRequest[field as keyof typeof partialRequest])) &&
+        partialRequest[field as keyof typeof partialRequest] !== null,
+    ) &&
+    (requireAmount
+      ? Boolean((partialRequest.srcTokenAmount ?? '').match(/^[1-9]\d*$/u))
+      : true)
+  );
+};
