+++ test/e2e/tests/bridge/constants.ts
@@ -1,4 +1,4 @@
-import { FeatureFlagResponse } from '../../../../ui/pages/bridge/types';
+import { FeatureFlagResponse } from '../../../../shared/types/bridge';
 
 export const DEFAULT_FEATURE_FLAGS_RESPONSE: FeatureFlagResponse = {
   'extension-config': {
