+++ ui/ducks/bridge/selectors.ts
@@ -42,7 +42,7 @@ import {
   QuoteMetadata,
   QuoteResponse,
   SortOrder,
-} from '../../pages/bridge/types';
+} from '../../../shared/types/bridge';
 import {
   calcAdjustedReturn,
   calcCost,
