+++ ui/pages/bridge/hooks/useHandleApprovalTx.ts
@@ -1,8 +1,15 @@
 import { TransactionType } from '@metamask/transaction-controller';
 import { Hex } from '@metamask/utils';
 import { BigNumber } from 'bignumber.js';
-import { TxData, QuoteResponse, FeeType } from '../types';
-import { isEthUsdt, getEthUsdtResetData } from '../bridge.util';
+import {
+  TxData,
+  QuoteResponse,
+  FeeType,
+} from '../../../../shared/types/bridge';
+import {
+  isEthUsdt,
+  getEthUsdtResetData,
+} from '../../../../shared/modules/bridge-utils/bridge.util';
 import { ETH_USDT_ADDRESS } from '../../../../shared/constants/bridge';
 import { getBridgeERC20Allowance } from '../../../ducks/bridge/actions';
 import { decimalToPrefixedHex } from '../../../../shared/modules/conversion.utils';
