+++ ui/pages/bridge/utils/quote.ts
@@ -1,7 +1,11 @@
 import { zeroAddress } from 'ethereumjs-util';
 import { BigNumber } from 'bignumber.js';
 import { calcTokenAmount } from '../../../../shared/lib/transactions-controller-utils';
-import { QuoteResponse, QuoteRequest, Quote, L1GasFees } from '../types';
+import {
+  QuoteResponse,
+  Quote,
+  L1GasFees,
+} from '../../../../shared/types/bridge';
 import {
   hexToDecimal,
   sumDecimals,
@@ -25,41 +29,6 @@ export const isQuoteExpired = (
 
 export const isNativeAddress = (address?: string | null) =>
   address === zeroAddress() || address === '' || !address;
-
-export const isValidQuoteRequest = (
-  partialRequest: Partial<QuoteRequest>,
-  requireAmount = true,
-): partialRequest is QuoteRequest => {
-  const STRING_FIELDS = ['srcTokenAddress', 'destTokenAddress'];
-  if (requireAmount) {
-    STRING_FIELDS.push('srcTokenAmount');
-  }
-  const NUMBER_FIELDS = ['srcChainId', 'destChainId', 'slippage'];
-
-  return (
-    STRING_FIELDS.every(
-      (field) =>
-        field in partialRequest &&
-        typeof partialRequest[field as keyof typeof partialRequest] ===
-          'string' &&
-        partialRequest[field as keyof typeof partialRequest] !== undefined &&
-        partialRequest[field as keyof typeof partialRequest] !== '' &&
-        partialRequest[field as keyof typeof partialRequest] !== null,
-    ) &&
-    NUMBER_FIELDS.every(
-      (field) =>
-        field in partialRequest &&
-        typeof partialRequest[field as keyof typeof partialRequest] ===
-          'number' &&
-        partialRequest[field as keyof typeof partialRequest] !== undefined &&
-        !isNaN(Number(partialRequest[field as keyof typeof partialRequest])) &&
-        partialRequest[field as keyof typeof partialRequest] !== null,
-    ) &&
-    (requireAmount
-      ? Boolean((partialRequest.srcTokenAmount ?? '').match(/^[1-9]\d*$/u))
-      : true)
-  );
-};
 
 export const calcToAmount = (
   { destTokenAmount, destAsset }: Quote,
